реализовал экран правил через AbstractView
